Added config.data_mapper.plugins configuration accessor
This allows to specify the needed plugins inside
the Rails.root/config/application.rb file like so

  config.data_mapper.plugins = %w(
	dm-validations
    dm-timestamps
    dm-constraints
    dm-types
  )

Note that dm-validations and dm-timestamps
are activated by default. Also note that setting
the plugins option in your application.rb file
currently resets those defaults, so if you want to
specify extra plugins, do list dm-validations and
dm-timestamps again. This actually isn't so bad
imho, because it gives a complete overview about
all plugins in use, and you don't have to remember
which plugins are loaded by default.

Note that requiring plugins in config/initializer
files is not yet tested and may be broken. This is
because dm generally needs its plugins loaded before
requiring the model files, in order to have things
like added class methods and such available.

I'm not yet sure how to influence the load order
of initializers defined in config/initializers,
so the only trusted way to activate them at the
appropriate time is to use the

  config.data_mapper.plugins = [ :'dm-plugin' ]

settings in config/application.rb