Support model, migration and observer generators
Looks like rails needs generators need to be in the

  ProjectName::Generators

namespace, which means that we need to use

  DataMapper::Generators

I'm not entirely happy with this and would rather
have used

  DataMapper::Rails::Generators or
  DattMapper::Generators::Rails

but I couldn't get the generators to run with either
of these namespacings. Maybe I overlooked something
and it's entirely possible after all.

However, here are the currently registered
generators as output by ./script/generate

 ...
 DataMapper:
  data_mapper:migration
  data_mapper:model
  data_mapper:observer
 ...

You can run them (and pass options) like so:

  ./script/generate data_mapper:model Person name:string
  ./script/generate data_mapper:migration Person
  ./script/generate data_mapper:observer person Person

I think the migration generator might take some
more options but I haven't yet explored them. Also
note that currently the migrations generator, for
some reason, isn't able to destroy generated
migrations using

  ./script/destroy migration Person

I'm not sure why this is, I always only get

  > remove migration.rb

which obviously doesn't exist. However, running
./script/generate migration Person works fine, and
puts the generated migration file below

  db/migrations/00X_person_migration.rb

Another thing to note is that the listed generators
still include the ones for ActiveRecord. I think
this is a bug in rails3, because I haven't even
required the 'active_record/railtie' in my sample
app.

Oh, and also I added 'dm-observer' to the list of
default plugins. This was mainly done to support
the observer generators. Maybe there is a way to
require dm-observer conditionally. This will need
more investigation.

Next thing to do is to move the rest of the code
to the DataMapper namespace, instead of Rails. I
think that actually fits better.