Cherry pick rails gems and drop rspec dependency for now
We don't need to require rails/all since we don't
want to install the activerecord gem. For this is
reason we don't want the rails metagem pulling in
activerecord for us.

This also drops the rspec-1.3 dependency for now.
Currently there are no specs anyway, and i won't
start writing any using 1.3 anymore. This will have
to wait till rspec-2.