Use require_dependency to preload models and lib
When STI is used for models the auto-require feature
of active_support may require a parent model when
requiring a child model before preload_models then
requires the same parent model again causing the
descendants of the parent to get corrupted and
relationship properties in the children end up not
getting defined correctly.

Maybe this could get fixed in dm-core? Sounds like
something weird is going on when STI related classes
are required/loaded twice? To be honest, I don't
exactly understand what's going on here, but Patrik
confirms that this patch fixes / works around the
STI issues.