Use a Benchmarking mixin instead of an Adapter wrapper
This patch completely drops the use of the adapter
cascade along with adapter wrappers.

This will hopefully work around the problems with
transactions and migrations, which both need real
AbstractAdapter descendants to work. The previous
adapter wrappers were'nt properly recognized which
basically rendered transactions and migrations
useless for the sake of being able to display some
benchmarking information in rails' server logs.

This patch should get around that by only extending
the adapter behavior with our benchmarking needs,
instead of wrapping it in a new object that's not
a descendant of AbstractAdapter.

Thx to Xavier Shay for the idea!

In the hopefully not so longterm, the benchmarking
mixin will go away anyway, because (at least) the
DO adapters will provide useful benchmarking info
by means of passing a struct containing the info
to a Connection#log(info) method that's defined in
ruby instead of only in the native driver
implementations (as is the case currently).