Make db:test:prepare a noop (at least for now)
Previously, running rake spec would always

  (1) create
  (2) auto_migrate!
  (3) seed

the database because db:test:prepare was coded to
do that. Then the spec_helper would have contained
another auto_migrate! command typically.

The proper behavior would probably be to at least
check for pending migrations, but this is not yet
implemented.

Making db:test:prepare a noop for now seems like a
reasonable way, because it gives full control over
auto_migrate! / auto_upgrade! to the spec_helper.rb
and this is how it's typically handled with DM.