Bumped dependencies to point to rails ~> 3.1.0.beta
For some reason, bundler always seems to hang with
almost 100% cpu at the following step:

  Fetching source index for http://rubygems.org/

This happens on both 1.8.7 and 1.9.2 when trying
to bundle install into a fresh rvm gemset.

I'm not sure why this is, if it's a bug with our
Gemfile or a bug in rails or bundler.

It's be cool if someone could try to reproduce so
that we can get down to the real error.